refactor(adex-cli): orderbook response processing moved (#1879)
This commit is part of adex-cli, it refactors the `orderbook` mod by moving it from the internal `response_handler` to its appropriate folder, enhancing code organization and clarity.
